requirements.txt
mysql-connector-python==8.2.0
python-dotenv==1.0.0
bcrypt==4.2.1
pytest==8.3.4